// generated/prisma/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* You can import this file directly.
*/

export const TwoFactorType = {
  SMS: 'SMS',
  TOTP: 'TOTP',
  EMAIL: 'EMAIL'
} as const

export type TwoFactorType = (typeof TwoFactorType)[keyof typeof TwoFactorType]


export const RoleType = {
  SUPERADMIN: 'SUPERADMIN',
  DIRECTEUR: 'DIRECTEUR',
  GERANT: 'GERANT',
  VENDEUR: 'VENDEUR',
  MAGASINIER: 'MAGASINIER'
} as const

export type RoleType = (typeof RoleType)[keyof typeof RoleType]


export const SubscriptionStatus = {
  TRIAL: 'TRIAL',
  ACTIVE: 'ACTIVE',
  SUSPENDED: 'SUSPENDED',
  CANCELLED: 'CANCELLED',
  EXPIRED: 'EXPIRED'
} as const

export type SubscriptionStatus = (typeof SubscriptionStatus)[keyof typeof SubscriptionStatus]


export const StockMovementType = {
  IN: 'IN',
  OUT: 'OUT',
  ADJUSTMENT: 'ADJUSTMENT',
  TRANSFER: 'TRANSFER',
  RETURN: 'RETURN'
} as const

export type StockMovementType = (typeof StockMovementType)[keyof typeof StockMovementType]


export const PaymentMethod = {
  CASH: 'CASH',
  CARD: 'CARD',
  MOBILE_MONEY: 'MOBILE_MONEY',
  BANK_TRANSFER: 'BANK_TRANSFER',
  CHEQUE: 'CHEQUE'
} as const

export type PaymentMethod = (typeof PaymentMethod)[keyof typeof PaymentMethod]